Make actual videos to make gifs from, keep actual videos

+++ lib/lolcommits/runner.rb
@@ -6,7 +6,7 @@ module Lolcommits
   class Runner
     attr_accessor :capture_delay, :capture_stealth, :capture_device, :message,
                   :sha, :snapshot_loc, :main_image, :config, :vcs_info,
-                  :capture_animate
+                  :capture_animate, :video_loc, :main_video
 
     def initialize(attributes = {})
       attributes.each do |attr, val|
@@ -49,12 +49,14 @@ module Lolcommits
       run_capture
 
       # check capture succeded, file must exist
-      if File.exist?(snapshot_loc)
-        ## resize snapshot first
-        resize_snapshot!
+      if File.exist?(snapshot_loc) || (capture_animated? && File.exist?(video_loc))
 
         # execute post_capture plugins, use to alter the capture
+        resize_snapshot! if File.exist?(snapshot_loc)
+                
         execute_plugins_for(:post_capture)
+
+        make_animated_gif if capture_animated?
 
         # execute capture_ready plugins, capture is ready for export/sharing
         execute_plugins_for(:capture_ready)
@@ -72,6 +74,8 @@ module Lolcommits
       puts '*** Preserving this moment in history.' unless capture_stealth
       self.snapshot_loc = config.raw_image(image_file_type)
       self.main_image   = config.main_image(sha, image_file_type)
+      self.main_video   = config.main_image(sha, 'mp4')
+      self.video_loc    = config.video_loc
 
       capturer = Platform.capturer_class(capture_animated?).new(
         capture_device: capture_device,
@@ -89,6 +93,58 @@ module Lolcommits
     end
 
     private
+
+    def capture_delay_string
+      " -ss #{capture_delay}" if capture_delay.to_i > 0
+    end
+
+    def frame_delay(fps, skip)
+      # calculate frame delay
+      delay = ((100.0 * skip) / fps.to_f).to_i
+      delay < 6 ? 6 : delay # hard limit for IE browsers
+    end
+
+    def video_fps(file)
+      # inspect fps of the captured video file (default to 29.97)
+      fps = system_call("ffmpeg -nostats -v quiet -i \"#{file}\" 2>&1 | sed -n \"s/.*, \\(.*\\) fp.*/\\1/p\"", true)
+      fps.to_i < 1 ? 29.97 : fps.to_f
+    end
+
+    def frame_skip(fps)
+      # of frames to skip depends on movie fps
+      case fps
+      when 0..15
+        2
+      when 16..28
+        3
+      else
+        4
+      end
+    end
+
+    def make_animated_gif
+      null_string = "/dev/null"
+      if Lolcommits::Platform.platform_windows?
+        null_string = "nul"
+      end
+      system_call "ffmpeg #{capture_delay_string} -nostats -v quiet -i \"#{main_video}\" -t #{capture_animate} \"#{config.frames_loc}/%09d.png\" > #{null_string}"
+
+      # use fps to set delay and number of frames to skip (for lower filesized gifs)
+      fps   = video_fps(main_video)
+      skip  = frame_skip(fps)
+      delay = frame_delay(fps, skip)
+      debug "Capturer: animated gif choosing every #{skip} frames with a frame delay of #{delay} (video fps: #{fps})"
+
+      # create the looping animated gif from frames (delete frame files except every #{skip} frame)
+      Dir["#{config.frames_loc}/*.png"].each do |frame_filename|
+          basename = File.basename(frame_filename)
+          frame_number = basename.split('.').first.to_i
+          File.delete(frame_filename) if frame_number % skip != 0
+      end
+
+      # convert to animated gif with delay and gif optimisation
+      system_call "convert -layers OptimizeTransparency -delay #{delay} -loop 0 \"#{config.frames_loc}/*.png\" -coalesce \"#{snapshot_loc}\""
+    end
 
     def enabled_plugins
       @enabled_plugins ||= config.plugin_manager.enabled_plugins_for(self)
@@ -115,10 +171,16 @@ module Lolcommits
       FileUtils.cp(snapshot_loc, main_image)
     end
 
+    def system_call(call_str, capture_output = false)
+      debug "Capturer: making system call for \n #{call_str}"
+      capture_output ? `#{call_str}` : system(call_str)
+    end
+
+
     def cleanup!
       debug 'Runner: running cleanup'
       # clean up the captured image and any other raw assets
-      FileUtils.rm(snapshot_loc)
+      FileUtils.rm(snapshot_loc) if File.exists?(snapshot_loc)
       FileUtils.rm_f(config.video_loc)
       FileUtils.rm_rf(config.frames_loc)
     end
